% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare.R
\name{Prepare}
\alias{Prepare}
\title{Prepare Dataset}
\usage{
Prepare(df, source, drop_na = FALSE, make_wide = FALSE, var_name = NULL)
}
\arguments{
\item{df}{DataFrame to be cleaned.}

\item{source}{The source of the data ('wb' for World Bank, 'un' for United Nations).}

\item{drop_na}{Logical; if TRUE, NA values will be dropped.}

\item{make_wide}{Logical; if TRUE, the dataset will be converted to wide format.}

\item{var_name}{Optional; specify a name for the variable column.}
}
\value{
A cleaned and optionally transformed DataFrame.
}
\description{
This function prepares datasets from specified sources for analysis.
}
\examples{
prepared_data <- Prepare(world_bank_example, source = 'wb')
}
