% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare.R
\name{std_country}
\alias{std_country}
\title{Standardize Country Names}
\usage{
std_country(df, column_name)
}
\arguments{
\item{df}{A dataframe containing a column with country names or codes.}

\item{column_name}{The name of the column in `df` that contains the country names or codes to be standardized.}
}
\value{
A dataframe with the specified country column replaced by a standardized "Country" column.
}
\description{
This function standardizes the names of countries in a given column of a dataframe
using the `countrycode` package. It maps country names or codes to their standardized
ISO 3-letter codes, removes the original country column, and renames the new column to "Country".
column replaced by a standardized "Country" column. Rows with countries that cannot be converted will be removed.
}
\examples{
data <- data.frame(country = c("USA", "France", "GBR"))
standardized_data <- std_country(data, "country")
}
