% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare.R
\name{UN_Clean}
\alias{UN_Clean}
\title{Clean United Nations Data}
\usage{
UN_Clean(dataset, drop_na = FALSE, make_wide = FALSE)
}
\arguments{
\item{dataset}{The dataset to be cleaned, typically obtained from United Nations sources.}

\item{drop_na}{Logical; if TRUE, rows with NA values will be dropped.}

\item{make_wide}{Logical; if TRUE, the dataset will be pivoted to wide format, with years as column headers.}
}
\value{
A cleaned DataFrame, formatted according to specified parameters.
}
\description{
This function cleans and formats dataset specifically from the United Nations. It standardizes
column names and formats, removes unnecessary columns, and handles data encoding and value
transformations.
}
\examples{
un_data <- UN_Clean(united_nations_example, drop_na = TRUE, make_wide = TRUE)
}
