% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/united_nations_example.R
\docType{data}
\name{united_nations_example}
\alias{united_nations_example}
\title{united_nations_example}
\format{
A data frame with 59 rows and 7 variables:
\describe{
  \item{Region/Country/Area}{Factor. Number to identify Country.}
  \item{Region/Country/Area Name}{Character. Name of Region/Country/Area.}
  \item{Year}{Integer. Year.}
  \item{Series}{Character. Some indicator.}
  \item{Value}{Numeric. Indicator value.}
  \item{Footnotes}{Character. Relevant Footnote or additional information.}
  \item{Source}{Character. Source of indicator value.}
}
}
\source{
United Nations Data: [https://data.un.org/_Docs/SYB/CSV/SYB66_1_202310_Population,%20Surface%20Area%20and%20Density.csv]
}
\usage{
united_nations_example
}
\description{
A dataset containing sample data from the United Nations, used for demonstrating
the functionality of the Prepare package.
}
\keyword{datasets}
