% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare.R
\name{WB_Clean}
\alias{WB_Clean}
\title{Clean World Bank Data}
\usage{
WB_Clean(dataset, drop_na = FALSE, make_wide = FALSE, var_name = NULL)
}
\arguments{
\item{dataset}{The dataset to be cleaned.}

\item{drop_na}{Logical; if TRUE, NA values will be dropped.}

\item{make_wide}{Logical; if TRUE, the dataset will be converted to wide format.}

\item{var_name}{Optional; specify a name for the variable column if `make_wide` is FALSE.}
}
\value{
A cleaned and optionally transformed DataFrame.
}
\description{
This function cleans and formats dataset specifically from the World Bank. It standardizes
column names and formats, removes unnecessary columns, and handles data encoding and value
transformations.
}
\examples{
wb_data <- WB_Clean(world_bank_example, drop_na = TRUE, make_wide = FALSE)
}
