% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_bank_example.R
\docType{data}
\name{world_bank_example}
\alias{world_bank_example}
\title{world_bank_example}
\format{
A data frame with 99 rows and 65 variables:
\describe{
  \item{Country Name}{Character. Number to identify Country.}
  \item{Country Code}{Character. Three letter country identifier.}
  \item{Indicator Name}{Character. Name of indicator.}
  \item{Indicator Code}{Character. UN Code for indicator.}
  \item{1960}{Numeric. Value of indicator in 1960.}
  Note: Additional columns 1961-2022 are in dataset but not described here. Columns for each year from 1961 to 2022 contain numeric values representing the value of the indicator for that year
}
}
\source{
World Bank Data: [https://data.worldbank.org/indicator/EN.ATM.CO2E.PC?view=chart]
}
\usage{
world_bank_example
}
\description{
A dataset containing sample data from the World Bank, used for demonstrating
the functionality of the Prepare package.
}
\keyword{datasets}
